+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `Assent.Strategy.OIDC.callback/2` now requires `:session_params` to be set in the config
 * `Assent.Strategy.OAuth2` now uses constant time comparison for state
 * `Assent.Strategy.OIDC` now uses constant time comparison for nonce
+* `Assent.Strategy.Httpc.request/5` bug fixed for certificates that has wildcard domain with SAN extension
 
 ## v0.1.10 (2020-04-23)